renamed canvas.js to pixelPainter.js
