+++ public/scripts/picturePainter.js
